Canal Cruiser Pause Panel
Got the clicking on the pause icon to work for pausing the Canal Cruiser minigame.

I also put in pressing the space bar to pause as well (just as a practice, in case you want to change to using a different key)

Also got the clicking on "Resume" and "Quit To Menu" buttons working.